+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change log
 
+-Simple Stack 1.1.0 (2017-02-18)
+--------------------------------
+- BREAKING CHANGE: `Backstack`'s APIs return `Object` instead of `Parcelable` (that includes `StateChange`, initial keys, `HistoryBuilder`, etc.)
+
+- ENHANCEMENT: Added `KeyParceler` interface to allow defining custom strategy in order to turn keys into `Parcelable` (for example, using `Parceler` library instead)
+
 -Simple Stack 1.0.0 (2017-02-15)
 --------------------------------
 - RELEASE: 1.0.0!